- organized code into NavigationManager
