+++ src/Core/src/Handlers/NavigationPage/NavigationViewHandler.Android.cs
@@ -9,6 +9,8 @@ namespace Microsoft.Maui.Handlers
 	internal partial class NavigationViewHandler :
 		ViewHandler<INavigationView, NavigationLayout>
 	{
+		NavigationManager? _navigationManager;
+
 		protected override NavigationLayout CreateNativeView()
 		{
 			LayoutInflater? li = LayoutInflater.From(Context);
@@ -20,16 +22,22 @@ namespace Microsoft.Maui.Handlers
 			return view;
 		}
 
+		// this should move to a factory method
+		protected virtual NavigationManager CreateNavigationManager() =>
+			_navigationManager ??= new NavigationManager();
+
 		protected override void ConnectHandler(NavigationLayout nativeView)
 		{
+			_navigationManager = CreateNavigationManager();
+			nativeView.NavigationManager = _navigationManager;
 			base.ConnectHandler(nativeView);
-			NativeView.SetVirtualView(VirtualView);
+			_navigationManager.Connect(VirtualView, nativeView);
 		}
 
 		public static void RequestNavigation(NavigationViewHandler arg1, INavigationView arg2, object? arg3)
 		{
 			if (arg3 is MauiNavigationRequestedEventArgs ea)
-				arg1.NativeView.RequestNavigation(ea);
+				arg1._navigationManager?.RequestNavigation(ea);
 		}
 	}
 }
