+++ src/Core/src/Platform/Android/Navigation/NavigationLayout.cs
@@ -1,33 +1,23 @@
 ﻿using System;
-using System.Collections.Generic;
 using Android.Content;
-using Android.OS;
 using Android.Runtime;
 using Android.Util;
 using AndroidX.AppCompat.Widget;
 using AndroidX.CoordinatorLayout.Widget;
-using AndroidX.Navigation;
-using AndroidX.Navigation.Fragment;
 using Google.Android.Material.AppBar;
 
 namespace Microsoft.Maui
 {
-	public class NavigationLayout : CoordinatorLayout, NavController.IOnDestinationChangedListener
+	public class NavigationLayout : CoordinatorLayout
 	{
-		NavHostFragment? _navHost;
-		FragmentNavigator? _fragmentNavigator;
-		NavigationStackNavGraph? _navGraphDestination;
-		Toolbar? _toolbar;
-		AppBarLayout? _appBar;
-
-		internal NavigationStackNavGraph NavGraphDestination => _navGraphDestination ??
-			throw new InvalidOperationException($"NavigationStackNavGraph cannot be null");
-
-		internal IView? VirtualView { get; private set; }
-		internal INavigationView? NavigationView { get; private set; }
-
-		public IMauiContext MauiContext => VirtualView?.Handler?.MauiContext ??
-			throw new InvalidOperationException($"MauiContext cannot be null");
+		// I'd prefer to not need this here but I'm not sure how else
+		// to get this class to the NavigationViewFragment
+		NavigationManager? _navigationManager;
+		public NavigationManager NavigationManager
+		{
+			get => _navigationManager ?? throw new InvalidOperationException($"NavigationManager cannot be null");
+			internal set => _navigationManager = value;
+		}
 
 #pragma warning disable CS0618 //FIXME: [Preserve] is obsolete
 		[Preserve(Conditional = true)]
@@ -50,153 +40,5 @@ namespace Microsoft.Maui
 		{
 		}
 #pragma warning restore CS0618 //FIXME: [Preserve] is obsolete
-
-		internal NavHostFragment NavHost
-		{
-			get => _navHost ?? throw new InvalidOperationException($"NavHost cannot be null");
-			set => _navHost = value;
-		}
-
-		internal FragmentNavigator FragmentNavigator
-		{
-			get => _fragmentNavigator ?? throw new InvalidOperationException($"FragmentNavigator cannot be null");
-			set => _fragmentNavigator = value;
-		}
-
-		internal Toolbar Toolbar
-		{
-			get => _toolbar ?? throw new InvalidOperationException($"ToolBar cannot be null");
-			set => _toolbar = value;
-		}
-
-		internal AppBarLayout AppBar
-		{
-			get => _appBar ?? throw new InvalidOperationException($"AppBar cannot be null");
-			set => _appBar = value;
-		}
-
-		public virtual void SetVirtualView(IView navigationView)
-		{
-			_toolbar = FindViewById<Toolbar>(Resource.Id.maui_toolbar);
-			_appBar = FindViewById<AppBarLayout>(Resource.Id.appbar);
-
-			VirtualView = navigationView;
-			NavigationView = (INavigationView)navigationView;
-
-			var fragmentManager = Context?.GetFragmentManager();
-			_ = fragmentManager ?? throw new InvalidOperationException($"GetFragmentManager returned null");
-			_ = NavigationView ?? throw new InvalidOperationException($"VirtualView cannot be null");
-
-			NavHost = (NavHostFragment)
-				fragmentManager.FindFragmentById(Resource.Id.nav_host);
-
-			FragmentNavigator =
-				(FragmentNavigator)NavHost
-					.NavController
-					.NavigatorProvider
-					.GetNavigator(Java.Lang.Class.FromType(typeof(FragmentNavigator)));
-		}
-
-		// Fragments are always destroyed if they aren't visible
-		// The Handler/NativeView associated with the visible IView remain intact
-		// The performance hit of destorying/recreating fragments should be negligible
-		// Hopefully this behavior survives implementation
-		// This will need to be tested with Maps and WebViews to make sure they behave efficiently
-		// being removed and then added back to a different Fragment
-		// 
-		// I'm firing NavigationFinished from here instead of FragmentAnimationFinished because
-		// this event appears to fire slightly after `FragmentAnimationFinished` and it also fires
-		// if we aren't using animations
-		private void OnNavigationViewFragmentDestroyed(AndroidX.Fragment.App.FragmentManager fm, NavigationViewFragment navHostPageFragment)
-		{
-			_ = NavigationView ?? throw new InvalidOperationException($"NavigationView cannot be null");
-
-			if (NavGraphDestination.IsNavigating)
-			{
-				NavGraphDestination.NavigationFinished(this.NavigationView);
-			}
-		}
-
-		protected virtual void OnNavigationViewFragmentResumed(AndroidX.Fragment.App.FragmentManager fm, NavigationViewFragment navHostPageFragment)
-		{
-			if (NavGraphDestination.IsInitialNavigation)
-			{
-				NavGraphDestination.NavigationFinished(this.NavigationView);
-			}
-		}
-
-		public virtual void RequestNavigation(MauiNavigationRequestedEventArgs e)
-		{
-			if (_navGraphDestination == null)
-			{
-				var navGraphNavigator =
-				   (NavGraphNavigator)NavHost
-					   .NavController
-					   .NavigatorProvider
-					   .GetNavigator(Java.Lang.Class.FromType(typeof(NavGraphNavigator)));
-
-				_navGraphDestination = new NavigationStackNavGraph(navGraphNavigator);
-
-				NavHost.NavController.AddOnDestinationChangedListener(this);
-				NavHost.ChildFragmentManager.RegisterFragmentLifecycleCallbacks(new FragmentLifecycleCallback(this), false);
-			}
-
-			NavGraphDestination.ApplyNavigationRequest(e, this);
-		}
-
-		internal void BackButtonPressed()
-		{
-			_ = NavigationView ?? throw new InvalidOperationException($"NavigationView cannot be null");
-
-			var graph = (NavigationStackNavGraph)NavHost.NavController.Graph;
-			var stack = new List<IView>(graph.NavigationStack);
-			stack.RemoveAt(stack.Count - 1);
-			graph.ApplyNavigationRequest(new MauiNavigationRequestedEventArgs(stack, true) , this);
-		}
-
-		#region IOnDestinationChangedListener
-
-		protected virtual void OnDestinationChanged(NavController navController, NavDestination navDestination, Bundle bundle)
-		{
-		}
-
-		void NavController.IOnDestinationChangedListener.OnDestinationChanged(
-			NavController p0, NavDestination p1, Bundle p2)
-		{
-			OnDestinationChanged(p0, p1, p2);
-		}
-		#endregion
-
-		class FragmentLifecycleCallback : AndroidX.Fragment.App.FragmentManager.FragmentLifecycleCallbacks
-		{
-			NavigationLayout _navigationLayout;
-
-			public FragmentLifecycleCallback(NavigationLayout navigationLayout)
-			{
-				_navigationLayout = navigationLayout;
-			}
-
-			public override void OnFragmentResumed(AndroidX.Fragment.App.FragmentManager fm, AndroidX.Fragment.App.Fragment f)
-			{
-				if (f is NavigationViewFragment pf)
-					_navigationLayout.OnNavigationViewFragmentResumed(fm, pf);
-			}
-
-			public override void OnFragmentAttached(AndroidX.Fragment.App.FragmentManager fm, AndroidX.Fragment.App.Fragment f, Context context)
-			{
-				base.OnFragmentAttached(fm, f, context);
-			}
-
-			public override void OnFragmentViewDestroyed(
-				AndroidX.Fragment.App.FragmentManager fm,
-				AndroidX.Fragment.App.Fragment f)
-			{
-				if (f is NavigationViewFragment pf)
-					_navigationLayout.OnNavigationViewFragmentDestroyed(fm, pf);
-
-				base.OnFragmentViewDestroyed(fm, f);
-			}
-		}
-
 	}
 }
